+++ packages/core/src/modules/application/application.core.ts
@@ -6,11 +6,11 @@ import { type DB, db } from '@oyster/db';
 import { type Application, OtherDemographic } from '@oyster/types';
 import { id, run } from '@oyster/utils';
 
+import { job } from '@/infrastructure/bull/bull';
 import {
   ApplicationBullJob,
   type GetBullJobData,
 } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import {
   type ApplicationRejectionReason,
