+++ packages/core/src/modules/authentication/use-cases/send-one-time-code.ts
@@ -3,7 +3,7 @@ import { match } from 'ts-pattern';
 import { db } from '@oyster/db';
 import { id } from '@oyster/utils';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { sendEmail } from '@/modules/notification/use-cases/send-email';
 import { IS_DEVELOPMENT } from '@/shared/env';
 import type {
