+++ packages/core/src/modules/authentication/use-cases/verify-one-time-code.ts
@@ -2,7 +2,7 @@ import { match } from 'ts-pattern';
 
 import { db } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { IS_DEVELOPMENT } from '@/shared/env';
 import {
   type OneTimeCodePurpose,
