+++ packages/core/src/modules/education/use-cases/add-education.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 import { id } from '@oyster/utils';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { checkMostRecentEducation } from './check-most-recent-education';
 import { type AddEducationInput } from '../education.types';
