+++ packages/core/src/modules/education/use-cases/delete-education.ts
@@ -1,6 +1,6 @@
 import { db } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { checkMostRecentEducation } from './check-most-recent-education';
 import { type Education } from '../education.types';
