+++ packages/core/src/modules/employment/use-cases/add-company-review.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 import { id } from '@oyster/utils';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { type AddCompanyReviewInput } from '../employment.types';
 
 export async function addCompanyReview({
