+++ packages/core/src/modules/employment/use-cases/add-work-experience.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 import { id } from '@oyster/utils';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { saveCompanyIfNecessary } from './save-company-if-necessary';
 import { type AddWorkExperienceInput } from '../employment.types';
