+++ packages/core/src/modules/employment/use-cases/delete-work-experience.ts
@@ -1,6 +1,6 @@
 import { db } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { type DeleteWorkExperienceInput } from '../employment.types';
 
 /**
