+++ packages/core/src/modules/event/use-cases/register-for-event.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { registerForAirmeetEvent } from '../airmeet-event.service';
 
 export async function registerForEvent({
