+++ packages/core/src/modules/feed/feed.ts
@@ -5,11 +5,11 @@ import { match } from 'ts-pattern';
 
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import {
   FeedBullJob,
   type GetBullJobData,
 } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import { ENV } from '@/shared/env';
