+++ packages/core/src/modules/member/events/member-activated.ts
@@ -2,8 +2,8 @@ import dayjs from 'dayjs';
 
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { createGiftCard } from '@/modules/shopify';
 
 export async function onMemberActivated({
