+++ packages/core/src/modules/member/events/member-email-added.ts
@@ -1,5 +1,5 @@
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 
 export async function onEmailAdded({
   email,
