+++ packages/core/src/modules/member/use-cases/activate-member.ts
@@ -1,6 +1,6 @@
 import { db } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 
 export async function activateMember(id: string) {
   await db
