+++ packages/core/src/modules/member/use-cases/backfill-engagement-records.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { getSlackUserByEmail } from '@/modules/slack/services/slack-user.service';
 
 type BackfillEngagementRecordsInput =
