+++ packages/core/src/modules/member/use-cases/change-primary-email.ts
@@ -1,6 +1,6 @@
 import { db } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import { type ChangePrimaryEmailInput } from '@/modules/member/member.types';
 
 /**
