+++ packages/core/src/modules/member/use-cases/join-member-directory.ts
@@ -2,7 +2,7 @@ import { type Transaction } from 'kysely';
 
 import { type DB } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 
 export async function joinMemberDirectory(
   trx: Transaction<DB>,
