+++ packages/core/src/modules/member/use-cases/send-birthday-notification.ts
@@ -2,8 +2,8 @@ import { sql } from 'kysely';
 
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { ENV } from '@/shared/env';
 
 export async function sendBirthdayNotification(
