+++ packages/core/src/modules/member/use-cases/update-member-email.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 import { type Student } from '@oyster/types';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 
 type UpdateMemberEmailInput = Pick<Student, 'email' | 'id'>;
