+++ packages/core/src/modules/onboarding-session/events/onboarding-session-attended.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 
 export async function onOnboardingSessionAttended({
   onboardingSessionId,
