+++ packages/core/src/modules/referral/referral.core.ts
@@ -3,7 +3,7 @@ import { type SelectExpression } from 'kysely';
 import { type DB, db } from '@oyster/db';
 import { id } from '@oyster/utils';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 import {
   type ReferFriendInput,
   ReferralStatus,
