+++ packages/core/src/modules/resume-review/resume-review.ts
@@ -6,11 +6,11 @@ import { z } from 'zod';
 import { db, relativeTime } from '@oyster/db';
 import { id } from '@oyster/utils';
 
+import { job } from '@/infrastructure/bull/bull';
 import {
   type GetBullJobData,
   ResumeReviewBullJob,
 } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { registerWorker } from '@/infrastructure/bull/use-cases/register-worker';
 import { getChatCompletion } from '@/modules/ai/ai';
 import { track } from '@/modules/mixpanel';
