+++ packages/core/src/modules/slack/events/slack-profile-picture-changed.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 
 export async function onSlackProfilePictureChanged({
   profilePicture,
