+++ packages/core/src/modules/slack/use-cases/delete-slack-message.ts
@@ -1,7 +1,7 @@
 import { db } from '@oyster/db';
 
+import { job } from '@/infrastructure/bull/bull';
 import { type GetBullJobData } from '@/infrastructure/bull/bull.types';
-import { job } from '@/infrastructure/bull/use-cases/job';
 import { retryWithBackoff } from '@/shared/utils/core.utils';
 
 export async function deleteSlackMessage({
