+++ packages/core/src/modules/slack/use-cases/send-stackbot-mass-notification.ts
@@ -1,6 +1,6 @@
 import { db } from '@oyster/db';
 
-import { job } from '@/infrastructure/bull/use-cases/job';
+import { job } from '@/infrastructure/bull/bull';
 
 export async function sendStackBotMassNotification(message: string) {
   const members = await db
